slice(blorktools): drag target migrated
